remove remote config source code (#891)
