+++ src/config/local_app_dependency_test.go
@@ -89,29 +89,6 @@ var _ = Describe("LocalAppDependency", func() {
 		})
 	})
 
-	var _ = Describe("exocom prod dependency", func() {
-		var exocomProd config.RemoteAppDependency
-		var _ = BeforeEach(func() {
-			for dependencyName, dependency := range appContext.Config.Remote.Dependencies {
-				if dependencyName == "exocom" {
-					exocomProd = config.NewRemoteAppDependency(dependencyName, dependency, appContext)
-					break
-				}
-			}
-		})
-
-		var _ = Describe("GetDeploymentConfig", func() {
-			It("should return the correct deployment config for exocom", func() {
-				actual, err := exocomProd.GetDeploymentConfig()
-				Expect(err).NotTo(HaveOccurred())
-				Expect(actual).To(Equal(map[string]string{
-					"version": "0.27.0",
-					"dnsName": "originate.com",
-				}))
-			})
-		})
-	})
-
 	var _ = Describe("generic dependency", func() {
 		var mongo *config.LocalAppDependency
