--- src/config/remote_app_dependency.go
@@ -1,23 +1,0 @@
-package config
-
-import (
-	"github.com/Originate/exosphere/src/types"
-	"github.com/Originate/exosphere/src/types/context"
-)
-
-// RemoteAppDependency contains methods that return config information about a dependency
-type RemoteAppDependency interface {
-	GetDeploymentConfig() (map[string]string, error)
-}
-
-// NewRemoteAppDependency returns an AppProductionDependency
-func NewRemoteAppDependency(dependencyName string, dependency types.RemoteDependency, appContext *context.AppContext) RemoteAppDependency {
-	switch dependency.Type {
-	case "exocom":
-		return &remoteExocomDependency{dependency, appContext}
-	case "rds":
-		return &remoteRdsDependency{dependencyName, dependency, appContext}
-	default:
-		return &remoteGenericDependency{dependency}
-	}
-}
